Control.Applicative.Free.Zero

Description

An applicative with left zero f is an Applicative with polymorphic zero action which stops all actions right to zero.

-- Type of zero action
zero :: forall a. f a

-- Left zero law
zero <*> x === zero

This module provides the free construction of them in a way similar to the free monad or the free applicative.

Documentation

data Free (f :: Type -> Type) a where

Free (applicative with left zero).

Constructors

Pure :: forall a (f :: Type -> Type). a -> Free f a

Zero :: forall (f :: Type -> Type) a. Free f a

Ap :: forall (f :: Type -> Type) a1 a. f a1 -> Free f (a1 -> a) -> Free f a

hoistFree :: (forall x. f x -> g x) -> Free f a -> Free g a

trap :: forall (f :: Type -> Type) a. Free f a -> Free f a -> Free f a

Recovery from Zero.

trap x y first look at x if it ends with Zero constructor. If x ends with Pure, return x itself. Otherwise, it replaces Zero in x with y.

trap (Ap f1 (Ap f2 ... Zero)) y === Ap f1 (Ap f2 ... y)

Interpreting

interpret :: Applicative f => (forall r. f r) -> Free f a -> f a

interpretAlternative :: Alternative f => Free f a -> f a

interpretMaybeT :: Monad f => Free f a -> f (Maybe a)