Control.Comonad.Travel

Description

Implementation of "Directed Containers as Categories," D. Ahman and T. Uustalu, EPTCS 207, 2016, pp. 89-98

https://arxiv.org/abs/1604.01187

Travel: Comonad from a Category

data Travel (cat :: k -> k -> Type) r where

Make Polynomial Comonad from Category.

Travel cat is a polynomial functor for any type constructor cat.

Travel cat r ~ ∑a. ((∑b. cat a b) -> r)

Travel cat is a Comonad if cat is a Category.

Constructors

MkTravel :: forall k (cat :: k -> k -> Type) r (a :: k). Sing a -> (Sigma k (TyCon (cat a)) -> r) -> Travel cat r

data TravelTag (cat :: k -> k -> Type) x where

Constructors

TravelFrom :: forall k (cat :: k -> k -> Type) (a :: k). Sing a -> TravelTag cat (Sigma k (TyCon (cat a)))

truncateTravel :: forall k (cat :: k -> k -> Type) r. SingKind k => Store (Demote k) r -> Travel cat r

Constructs Travel cat from Store (Demote k), which can be thought of as a Travel cat but its arrows are forgotten.