Control.Comonad.Travel.Boolean

Description

A small but non-trivial example of category - comonad correspondence

Documentation

data H r

Constructors

H1 r

H2 r r

data Implies (x :: Bool) (y :: Bool) where

Boolean algebra {False, True} can be seen as a category:

False ---> True
  ↺         ↺

Note: Roughly speaking, there are only three values in types Implies b1 b2, where b1, b2 are chosen from either False or True.

ImpId :: Implies False False
ImpId :: Implies True True
ImpFT :: Implies False True

Constructors

ImpId :: forall (x :: Bool). Implies x x

ImpFT :: Implies 'False 'True

toH :: Travel Implies r -> H r

Travel Implies is isomorphic to H, including its Comonad instance.

fromH :: H r -> Travel Implies r