Control.Monad.Flow

Flow: Monad from a Category

type Flow (cat :: k -> k -> Type) = FlowT cat Identity

Flow can be thought of as a generalization of the State Monad.

A value of State s x describes:

For each state s1 :: s, a next state s2 :: s

For each state s1 :: s, a value x1 :: x

A value of Flow cat describes:

For each object s1 :: k, an arrow f :: cat s1 s2 starting from s1 and going to a next object s2

For each object s1 :: k, a value x1 :: x

mkFlow :: forall k (cat :: k -> k -> Type) x. (forall (a :: k). Sing a -> (x, Sigma k (TyCon (cat a)))) -> Flow cat x

truncateFlow :: forall k (cat :: k -> k -> Type) x. SingKind k => Flow cat x -> State (Demote k) x

Forgets arrows of the underlying category from Flow cat, converting it to a simple State monad on objects of cat.

toCoTravel :: forall k (cat :: k -> k -> Type) x. Category cat => Flow cat x -> Co (Travel cat) x

Co (Travel cat) is isomorphic to Flow cat

fromCoTravel :: forall k (cat :: k -> k -> Type) x. Category cat => Co (Travel cat) x -> Flow cat x

Co (Travel cat) is isomorphic to Flow cat

Transformer version

newtype FlowT (cat :: k -> k -> Type) (m :: Type -> Type) x

Transformer version of Flow.

Constructors

MkFlowT

Fields

runFlowT :: forall (a :: k). Sing a -> m (x, Sigma k (TyCon (cat a)))

truncateFlowT :: forall k (cat :: k -> k -> Type) (m :: Type -> Type) x. (SingKind k, Monad m) => FlowT cat m x -> StateT (Demote k) m x

Forgets arrows of the underlying category from FlowT cat m, converting it to a simple StateT on objects of cat.