Data.Bifunctor.AlternatingList

Documentation

data AltList a b

  AltList a b = a | (a,b,a) | (a,b,a,b,a) | ...
  

Constructors

Last a

Next a b (AltList a b)

head :: AltList a b -> a

tail :: AltList a b -> Maybe (b, AltList a b)

separate :: AltList a b -> (NonEmpty a, [b])

foldMapBoth :: (Semigroup ra, Monoid rb) => (a -> ra) -> (b -> rb) -> AltList a b -> (ra, rb)

toNonEmptyBy :: (a -> c) -> (b -> c) -> AltList a b -> NonEmpty c

glue :: AltList a b -> b -> AltList a b -> AltList a b

glue xs b ys

xs = [ax, bx, ax, bx, ..., ax]
ys = [ay, by, ay, by, ..., ay]
glue xs b0 ys = [ax, bx, ax, bx, ..., ax, b0, ay, by, ..., by, ay]

flatMap :: (a -> AltList b c) -> AltList a c -> AltList b c

reverse :: AltList a b -> AltList a b

unfoldr :: (s -> Either a (a, b, s)) -> s -> AltList a b

unfoldl :: (s -> Either a (s, b, a)) -> s -> AltList a b

unfoldTree :: (s -> Either a (s, b, s)) -> s -> AltList a b

fromEithers :: Monoid a => [Either a b] -> AltList a b

fromEithers' :: Monoid a => [Either a b] -> AltList a b

bizip :: AltList a b -> AltList a' b' -> AltList (a, a') (b, b')

bizipWith :: (a -> a' -> a'') -> (b -> b' -> b'') -> AltList a b -> AltList a' b' -> AltList a'' b''

(TODO: documentation)

AltList could be an instance of Biapplicative using bizipWith as biliftA2, but that instance is not provided here.

This is because to avoid confusion between Applicative instance, which works like how 'Applicative []' instance work.

birepeat :: a -> b -> AltList a b

unzip :: AltList (a, a') (b, b') -> (AltList a b, AltList a' b')