Data.DoubleMonoid.Class

Documentation

class DoubleMonoid a where

DoubleMonoid is a type with two independent monoid structures.

The two monoids are named "additive" (zero, /+/) and "multiplicative" (one, /*/) respectively. But unlike Num or (semi)ring structures, DoubleMonoid assume no relation between the two monoids like distributivity.

Laws

Monoid laws of "additive" set of operators:

(x /+/ y) /+/ z === x /+/ (y /+/ z)
zero /+/ x === x
x /+/ zero === x

Monoid laws of "multiplicative" set of operators:

(x /*/ y) /*/ z === x /*/ (y /*/ z)
one /*/ x === x
x /*/ one === x

Relation to other known algebraic structures

Of course, DoubleMonoid can be seen as Monoid in two ways. Other than that, near-semiring (TODO:url) is a subclass of DoubleMonoid which additionally satisfy "left distribution law" (which confusingly have another name "right distributivity").

-- in near-semiring
zero /*/ x  === zero
(x /+/ y) /*/ z === (x /*/ z) /+/ (y /*/ z)

Minimal complete definition

one, (/*/), zero, (/+/) | mprod, msum

Methods

one :: a

The unit of "multiplicative" monoid.

(/*/) :: a -> a -> a infixr 6

The binary operator of "multiplicative" monoid.

zero :: a

The unit of "additive" monoid.

(/+/) :: a -> a -> a infixr 5

The binary operator of "additive" monoid.

mprod :: [a] -> a

Fold a list using /*/ monoid.

msum :: [a] -> a

Fold a list using /+/ monoid.