Data.DoubleMonoid.LZLC.Free

Documentation

data Free a

The free DMLZLC.

Constructors

Lit a

Zero

One

SumOf (ZList' (SummandF a))

ProdOf (ZList' (FactorF a))

data SummandF a

Constructors

SummandLit a

SummandProd (ZList' (FactorF a))

injectSummand :: SummandF a -> Free a

data FactorF a

Constructors

FactorLit a

FactorSum (ZList' (SummandF a))

injectFactor :: FactorF a -> Free a

interpret :: DMLZLC b => (a -> b) -> Free a -> b

viewSum :: Free a -> ZList (SummandF a)

viewProd :: Free a -> ZList (FactorF a)

mprodZ :: DMLZ a => ZList a -> a

(one, /*/) is a monoid with one right zero element zero. This can be stated as mprodZ is a ZList algebra.

mprodZ . fmap mprodZ === mprodZ . join

When you override the default implementation, it must keep the relation to the other methods:

mprodZ Nend === one
mprodZ Zend === zero
mprodZ (Cons a Nend) === a

reexport

data ZList a

List but ends in two ways.

If a list (x :: ZList a) ends with Nend, concatenation x <> y appends y like usual list.

Cons a (Cons a' (... Nend)) <> y === Cons a (Cons a' (... y))

On the other hand, if x ends with Zend, concatenation from right x <> y does not change x.

Cons a (Cons a' (... Zend)) <> _ === Cons a (Cons a' (... Zend))

Constructors

Nend

Zend

Cons a (ZList a)

data ZList' a

ZList sans Nend, Zend, Cons a Nend

Constructors

Zxz a

ZLong a a (ZList a)