Data.Finitary.PolyRep

Description

A normalized polynomial representation for finitary polynomial functors.

Compared with Data.Finitary.TreeRep, this module forgets the syntactic tree structure of sums and products and represents a functor as

x^n1 + x^n2 + ... + x^nk

This is especially useful for Day convolution: monomials satisfy

Day x^m x^n ≅ x^(m*n)

which is implemented by DayPoly, fromDay, and toDay.

The list order in Poly is operationally significant in Haskell, but mathematically it should be regarded as a chosen ordering of summands.

Base Type and its algebra

type Poly = [Nat]

Finitary polynomial f(x) = x^e1 + x^e2 + ... + x^en represented as a list of exponents [e1, e2, ..., en]

zeroPoly :: Poly

addPoly :: Poly -> Poly -> Poly

onePoly :: Poly

multPoly :: Poly -> Poly -> Poly

parPoly :: Poly

Type-level Poly algebra

type AddPoly (r1 :: Poly) (r2 :: Poly) = r1 ++ r2

type family MultPoly (r1 :: Poly) (r2 :: Poly) :: Poly where ...

Equations

MultPoly ('[] :: [Nat]) _1 = '[] :: [Nat]

MultPoly (e ': es) r2 = AddPoly (MultPoly1 e r2) (MultPoly es r2)

type family DayPoly (r1 :: Poly) (r2 :: Poly) :: Poly where ...

Equations

DayPoly ('[] :: [Nat]) _1 = '[] :: [Nat]

DayPoly (e ': es) r2 = AddPoly (DayPoly1 e r2) (DayPoly es r2)

data SPoly (r :: Poly) where

Constructors

SNil :: SPoly ('[] :: [Nat])

SCons :: forall (e :: Nat) (es :: [Nat]). !(SNat e) -> !(SPoly es) -> SPoly (e ': es)

sAddPoly :: forall (r1 :: Poly) (r2 :: Poly). SPoly r1 -> SPoly r2 -> SPoly (AddPoly r1 r2)

(%++) :: forall (r1 :: Poly) (r2 :: Poly). SPoly r1 -> SPoly r2 -> SPoly (AddPoly r1 r2)

sMultPoly :: forall (r1 :: Poly) (r2 :: Poly). SPoly r1 -> SPoly r2 -> SPoly (MultPoly r1 r2)

sDayPoly :: forall (r1 :: Poly) (r2 :: Poly). SPoly r1 -> SPoly r2 -> SPoly (DayPoly r1 r2)

class KnownPoly (p :: Poly) where

Methods

sPoly :: SPoly p

withKnownPoly :: forall (r :: Poly) result. SPoly r -> (KnownPoly r => result) -> result

Evaluating Poly as a Haskell Functor

data Eval (r :: Poly) x where

Constructors

EHere :: forall (e :: Nat) x (es :: [Nat]). !(Finite e -> x) -> Eval (e ': es) x

EThere :: forall (es :: [Nat]) x (e :: Nat). !(Eval es x) -> Eval (e ': es) x

Correspondence between sums, products, and Day convolution of Poly and its evaluation

absurdEval :: Eval ('[] :: [Nat]) a -> b

unitEval :: Eval '[0] a

fromSum :: forall (r1 :: Poly) proxy (r2 :: Poly) x. SPoly r1 -> proxy r2 -> Either (Eval r1 x) (Eval r2 x) -> Eval (r1 ++ r2) x

inlEval :: forall (r1 :: Poly) proxy (r2 :: [Nat]) x. SPoly r1 -> proxy r2 -> Eval r1 x -> Eval (r1 ++ r2) x

inrEval :: forall (r1 :: Poly) proxy (r2 :: Poly) x. SPoly r1 -> proxy r2 -> Eval r2 x -> Eval (r1 ++ r2) x

toSum :: forall (r1 :: Poly) proxy (r2 :: [Nat]) x. SPoly r1 -> proxy r2 -> Eval (r1 ++ r2) x -> Either (Eval r1 x) (Eval r2 x)

fromProduct :: forall (r1 :: Poly) (r2 :: Poly) x. SPoly r1 -> SPoly r2 -> Eval r1 x -> Eval r2 x -> Eval (MultPoly r1 r2) x

toProduct :: forall (r1 :: Poly) (r2 :: Poly) x. SPoly r1 -> SPoly r2 -> Eval (MultPoly r1 r2) x -> (Eval r1 x, Eval r2 x)

fromDay :: forall (r1 :: Poly) (r2 :: Poly) x. SPoly r1 -> SPoly r2 -> Day (Eval r1) (Eval r2) x -> Eval (DayPoly r1 r2) x

toDay :: forall (r1 :: Poly) (r2 :: Poly) x. SPoly r1 -> SPoly r2 -> Eval (DayPoly r1 r2) x -> Day (Eval r1) (Eval r2) x

Profunctor traversal

ptraverseEval :: forall p (r :: Poly) a b. (Cartesian p, Cocartesian p) => SPoly r -> p a b -> p (Eval r a) (Eval r b)

Building bidirectional encodings as a Eval r with Profunctor

data Encoder a b s t where

Constructors

Encoder :: forall (r :: Poly) s a b t. !(SPoly r) -> (s -> Eval r a) -> (Eval r b -> t) -> Encoder a b s t

idEncoder :: Encoder a b a b

Encoder for the identity functor.

It can be used to construct an encoder for arbitrary PTraversable functor using

  ptraverse idEncoder :: PTraversable t => Encoder a b (t a) (t b)
  

.

Internal-only type families

(++) :: [a] -> [a] -> [a] infixr 5

(++) appends two lists, i.e.,

[x1, ..., xm] ++ [y1, ..., yn] == [x1, ..., xm, y1, ..., yn]
[x1, ..., xm] ++ [y1, ...] == [x1, ..., xm, y1, ...]

If the first list is not finite, the result is the first list.

Performance considerations

Expand

This function takes linear time in the number of elements of the first list. Thus it is better to associate repeated applications of (++) to the right (which is the default behaviour): xs ++ (ys ++ zs) or simply xs ++ ys ++ zs, but not (xs ++ ys) ++ zs. For the same reason concat = foldr (++) [] has linear performance, while foldl (++) [] is prone to quadratic slowdown

Examples

Expand

>>> [1, 2, 3] ++ [4, 5, 6]
[1,2,3,4,5,6]

>>> [] ++ [1, 2, 3]
[1,2,3]

>>> [3, 2, 1] ++ []
[3,2,1]

type family MultPoly1 (e :: Nat) (r2 :: Poly) :: Poly where ...

Equations

MultPoly1 _1 ('[] :: [Nat]) = '[] :: [Nat]

MultPoly1 e (f ': fs) = (e + f) ': MultPoly1 e fs

type family DayPoly1 (e :: Nat) (r2 :: Poly) :: Poly where ...

Equations

DayPoly1 _1 ('[] :: [Nat]) = '[] :: [Nat]

DayPoly1 e (f ': fs) = (e * f) ': DayPoly1 e fs