Data.Functor.Bicompose

Documentation

newtype Bicompose (f :: k2 -> Type) (g :: k0 -> k1) (h :: k1 -> k2) (a :: k0)

Both-side composition of Monad.

If both of f and g are Monad, Bicompose f g is an instance of FMonad in a similar way Compose and Precompose are.

Constructors

Bicompose

Fields

getBicompose :: f (h (g a))