Data.Functor.Polynomial.Tag

Description

This module provides a way to represent "tag" of various polynomial data types. For example, a GADT TagMaybe represents a functor of the same shape of Maybe.

data TagMaybe n where
   TagNothing :: TagMaybe Void -- ^ Nothing takes 0 parameter
   TagJust    :: TagMaybe ()   -- ^ Just takes 1 paramerer

The type class for tags

class HasFinitary (tag :: Type -> Type) where

Methods

withFinitary :: tag n -> (Finitary n => r) -> r

Tags

data TagMaybe n where

Tag of Maybe.

Constructors

TagNothing :: TagMaybe Void

TagJust :: TagMaybe ()

data TagList a where

Constructors

TagList :: forall (n :: Nat). KnownNat n => TagList (Finite Integer n)

type TagFn = (:~:) :: k -> k -> Type

TagFn n represents ((), const n :: () -> Type).

This is the tag of (->) r functor.

It's also the tag of Proxy and Identity, by being isomorphic to (->) Void and (->) () respectively.

TagFn Void   -- Tag of Proxy
TagFn ()     -- Tag of Identity

data TagV n

TagV n is an empty type for any n. It represents (∅, α = absurd :: ∅ -> Nat).

This is the tag of V1.

data TagK c n where

TagK c n represents (c, α), where α is a constant function to Void:

α = const Void :: c -> Type

This is the tag of Const.

Constructors

TagK :: forall c. c -> TagK c Void

type TagSum = (:+:) :: (k -> Type) -> (k -> Type) -> k -> Type

When t1, t2 represents (U, α1) and (V,α2) respectively,

α1 :: U -> Nat
α2 :: V -> Nat

TagSum t1 t2 represents (Either U V, β = either α1 α2).

β :: Either U V -> Nat
β (Left u)  = α1(u)
β (Right v) = α2(v)

This is the tag of f :+: g, when t1, t2 is the tag of f, g respectively.

data TagMul (f :: Type -> Type) (g :: Type -> Type) x where

When t1, t2 represents (U, α1) and (V,α2) respectively,

α1 :: U -> Nat
α2 :: V -> Nat

TagMul t1 t2 represents the direct product (U,V) and the function β on it:

β :: (U,V) -> Nat
β(u,v) = α1 u + α2 v

This is the tag of f :*: g when t1, t2 is the tag of f, g respectively.

Constructors

TagMul :: forall (f :: Type -> Type) x1 (g :: Type -> Type) y. !(f x1) -> !(g y) -> TagMul f g (Either x1 y)

data TagPow (n :: Natural) (t :: Type -> Type) xs where

When t represents (U,α :: U -> Type), TagPow n t xs represents α^n below:

α^n :: U^n -> Type
α^n(u1, u2, ..., u_n) = α u1 + α u2 + ... + α u_n

This is the tag of Pow n f when t is the tag of f.

Constructors

TagPow0 :: forall (t :: Type -> Type). TagPow 0 t Void

TagPowNZ :: forall (n :: Natural) (t :: Type -> Type) xs. TagPow' n t xs -> TagPow n t xs

geqPow :: forall (t :: Type -> Type) (n :: Natural) xs (n' :: Natural) xs'. GEq t => TagPow n t xs -> TagPow n' t xs' -> Maybe (n :~: n', xs :~: xs')

gcomparePow :: forall (t :: Type -> Type) (n :: Natural) xs (n' :: Natural) xs'. GCompare t => TagPow n t xs -> TagPow n' t xs' -> GOrdering '(n, xs) '(n', xs')

data TagComp (t :: Type -> Type) (u :: Type -> Type) n where

When t, u is the tag of f, g respectively, TagComp t u is the tag of f :.: g.

Constructors

TagComp :: forall (t :: Type -> Type) a (u :: Type -> Type) n. t a -> TagPow (Cardinality a) u n -> TagComp t u n

data TagDay (t :: Type -> Type) (u :: Type -> Type) n where

Constructors

TagDay :: forall (t :: Type -> Type) n1 (u :: Type -> Type) m. t n1 -> u m -> TagDay t u (n1, m)

Reexports

class GShow (t :: k -> Type) where

Show-like class for 1-type-parameter GADTs. GShow t => ... is equivalent to something like (forall a. Show (t a)) => .... The easiest way to create instances would probably be to write (or derive) an instance Show (T a), and then simply say:

instance GShow t where gshowsPrec = defaultGshowsPrec

Methods

gshowsPrec :: forall (a :: k). Int -> t a -> ShowS

class GEq (f :: k -> Type) where

A class for type-contexts which contain enough information to (at least in some cases) decide the equality of types occurring within them.

This class is sometimes confused with TestEquality from base. TestEquality only checks type equality.

Consider

>>> data Tag a where TagInt1 :: Tag Int; TagInt2 :: Tag Int

The correct TestEquality Tag instance is

>>> :{
instance TestEquality Tag where
    testEquality TagInt1 TagInt1 = Just Refl
    testEquality TagInt1 TagInt2 = Just Refl
    testEquality TagInt2 TagInt1 = Just Refl
    testEquality TagInt2 TagInt2 = Just Refl
:}

While we can define

instance GEq Tag where
   geq = testEquality

this will mean we probably want to have

instance Eq Tag where
   _ == _ = True

Note: In the future version of some package (to be released around GHC-9.6 / 9.8) the forall a. Eq (f a) constraint will be added as a constraint to GEq, with a law relating GEq and Eq:

geq x y = Just Refl   ⇒  x == y = True        ∀ (x :: f a) (y :: f b)
x == y                ≡  isJust (geq x y)     ∀ (x, y :: f a)

So, the more useful GEq Tag instance would differentiate between different constructors:

>>> :{
instance GEq Tag where
    geq TagInt1 TagInt1 = Just Refl
    geq TagInt1 TagInt2 = Nothing
    geq TagInt2 TagInt1 = Nothing
    geq TagInt2 TagInt2 = Just Refl
:}

which is consistent with a derived Eq instance for Tag

>>> deriving instance Eq (Tag a)

Note that even if a ~ b, the geq (x :: f a) (y :: f b) may be Nothing (when value terms are inequal).

The consistency of GEq and Eq is easy to check by exhaustion:

>>> let checkFwdGEq :: (forall a. Eq (f a), GEq f) => f a -> f b -> Bool; checkFwdGEq x y = case geq x y of Just Refl -> x == y; Nothing -> True
>>> (checkFwdGEq TagInt1 TagInt1, checkFwdGEq TagInt1 TagInt2, checkFwdGEq TagInt2 TagInt1, checkFwdGEq TagInt2 TagInt2)
(True,True,True,True)

>>> let checkBwdGEq :: (Eq (f a), GEq f) => f a -> f a -> Bool; checkBwdGEq x y = if x == y then isJust (geq x y) else isNothing (geq x y)
>>> (checkBwdGEq TagInt1 TagInt1, checkBwdGEq TagInt1 TagInt2, checkBwdGEq TagInt2 TagInt1, checkBwdGEq TagInt2 TagInt2)
(True,True,True,True)

Methods

geq :: forall (a :: k) (b :: k). f a -> f b -> Maybe (a :~: b)

Produce a witness of type-equality, if one exists.

A handy idiom for using this would be to pattern-bind in the Maybe monad, eg.:

extract :: GEq tag => tag a -> DSum tag -> Maybe a
extract t1 (t2 :=> x) = do
    Refl <- geq t1 t2
    return x

Or in a list comprehension:

extractMany :: GEq tag => tag a -> [DSum tag] -> [a]
extractMany t1 things = [ x | (t2 :=> x) <- things, Refl <- maybeToList (geq t1 t2)]

(Making use of the DSum type from Data.Dependent.Sum in both examples)

class GEq f => GCompare (f :: k -> Type) where

Type class for comparable GADT-like structures. When 2 things are equal, must return a witness that their parameter types are equal as well (GEQ).

Methods

gcompare :: forall (a :: k) (b :: k). f a -> f b -> GOrdering a b

data GOrdering (a :: k) (b :: k) where

A type for the result of comparing GADT constructors; the type parameters of the GADT values being compared are included so that in the case where they are equal their parameter types can be unified.

Constructors

GLT :: forall {k} (a :: k) (b :: k). GOrdering a b

GEQ :: forall {k} (a :: k). GOrdering a a

GGT :: forall {k} (a :: k) (b :: k). GOrdering a b