Data.Polynomial.Functor.Deriving

Documentation

newtype ViaPolynomial (e :: Poly -> Type -> Type) (f :: Type -> Type) a

Generic Deriving

Constructors

ViaPolynomial

Fields

unwrapViaPolynomial :: f a

newtype Zippy (p :: Poly) x

Any polynomial functor Ev p can have a "zippy" applicative instance.

Example

>>> type P = 'S ('T ('S ('T 'U))) -- P(x) = 1 + x + x^2
>>> pure 0 :: Zippy P Int
Zippy {runZippy = Go (0 ::: Go (0 ::: End))}
>>> fx = Zippy $ Go ("A" ::: Go ("a" ::: End)) :: Zippy P String
>>> fy = Zippy $ Go ("B" ::: Stop) :: Zippy P String
>>> fz = Zippy $ Stop :: Zippy P String
>>> (++) <$> fx <*> fy
Zippy {runZippy = Go ("AB" ::: Stop)}
>>> (++) <$> fx <*> fz
Zippy {runZippy = Stop}
>>> (++) <$> fy <*> fz
Zippy {runZippy = Stop}

Constructors

Zippy

Fields

runZippy :: Ev p x

newtype Aligney (p :: Poly) x

Any polynomial functor with no constant term, Ev (T p), can have an "aligney" applicative and monad instance.

Example

>>> type Q = 'T ('S ('T 'U)) -- Q(x) = x + x^2
>>> pure 0 :: Aligney Q Int
Aligney {runAligney = 0 ::: Stop}
>>> fx = Aligney $ "A" ::: Go ("a" ::: End) :: Aligney Q String
>>> fy = Aligney $ "B" ::: Go ("b" ::: End) :: Aligney Q String
>>> fz = Aligney $ "C" ::: Stop :: Aligney Q String
>>> (++) <$> fx <*> fy
Aligney {runAligney = "AB" ::: Go ("ab" ::: End)}
>>> (++) <$> fx <*> fz
Aligney {runAligney = "AC" ::: Go ("aC" ::: End)}
>>> (++) <$> fy <*> fz
Aligney {runAligney = "BC" ::: Go ("bC" ::: End)}
>>> fx >>= \x -> if x == "A" then fz else fy
Aligney {runAligney = "C" ::: Go ("b" ::: End)}

Constructors

Aligney

Fields

runAligney :: Ev p x