Data.Profunctor.Cartesian

Documentation

class Profunctor p => Cartesian (p :: Type -> Type -> Type) where

Minimal complete definition

proUnit, (proProduct | (***) | (&&&))

Methods

proUnit :: p a ()

Unit of the product.

The type of proUnit can be understood as proUnit' :: p () () treated to have more generally-typed p a () using lmap.

const () :: a -> ()
proUnit' :: p () ()
lmap (const ()) proUnit' :: p a ()

proProduct :: (a -> (a1, a2)) -> ((b1, b2) -> b) -> p a1 b1 -> p a2 b2 -> p a b

(***) :: p a b -> p a' b' -> p (a, a') (b, b')

Product of two profunctors

(&&&) :: p a b -> p a b' -> p a (b, b')

Alternative way to define the product (pronounced "fan-out")

proPower :: forall (n :: Nat) a b. KnownNat n => p a b -> p (Finite n -> a) (Finite n -> b)

Function from finite types can be constructed as iterated product.

There is a default implementaion, but it can be more efficient implementation.

prodDay :: forall (r :: Type -> Type -> Type) (p :: Type -> Type -> Type) (q :: Type -> Type -> Type). Cartesian r => (p :-> r) -> (q :-> r) -> Day (,) p q :-> r

class Profunctor p => Cocartesian (p :: Type -> Type -> Type) where

Minimal complete definition

proEmpty, (proSum | (+++) | (|||))

Methods

proEmpty :: p Void b

Unit of the sum.

The type of proEmpty can be understood as proEmpty' :: p Void Void treated to have more generally-typed p Void b using rmap.

absurd    :: Void -> b
proEmpty' :: p Void Void
rmap absurd proEmpty' :: p Void b

proSum :: (a -> Either a1 a2) -> (Either b1 b2 -> b) -> p a1 b1 -> p a2 b2 -> p a b

(+++) :: p a b -> p a' b' -> p (Either a a') (Either b b')

Sum of two profunctors

(|||) :: p a b -> p a' b -> p (Either a a') b

Alternative way to define the sum (pronounced "fan-in")

proTimes :: forall (n :: Nat) a b. KnownNat n => p a b -> p (Finite n, a) (Finite n, b)

Pairing with finite types can be constructed as iterated sum.

There is a default implementaion, but it can be more efficient implementation.

sumDay :: forall (r :: Type -> Type -> Type) (p :: Type -> Type -> Type) (q :: Type -> Type -> Type). Cocartesian r => (p :-> r) -> (q :-> r) -> Day Either p q :-> r

Cartesian p iff forall x. Applicative (p x)

pureDefault :: Cartesian p => b -> p a b

proUnitDefault :: Applicative (p a) => p a ()

liftA2Default :: Cartesian p => (a -> b -> c) -> p x a -> p x b -> p x c

fanoutDefault :: (Profunctor p, Applicative (p a)) => p a b1 -> p a b2 -> p a (b1, b2)

Utilities

describeFinite :: forall (n :: Nat) p. (KnownNat n, Cartesian p, Cocartesian p) => p (Finite n) (Finite n)

describeFiniteBits :: (FiniteBits a, Cartesian p, Cocartesian p) => p a a

Laws

Cartesian laws

Instances of Cartesian must satisfy the following equations.

Left unit of ***

dimap unitL' unitL (proUnit *** p) === p, where unitL' = ((),) :: a -> ((), a); unitL = snd :: ((),a) -> a

Right unit of ***

dimap unitR' unitR (p *** proUnit) === p, where unitR' = (,()) :: a -> (a,()); unitR = fst :: (a,()) -> a

Associativity of ***

dimap assoc unassoc ((p *** q) *** r) === p *** (q *** r)

Some instance of Cartesian optionally satisfy the following equation.

Commutativity of ***

dimap swap swap (p *** q) === q *** p

In this library, an instance of Cartesian is called commutative Cartesian if it satisfy the above commutativity law.

Instances with and without commutativity

TODO fill here: most instance is not commutative, Counting or Forget r for commutative monoid is (positive)

Cocartesian laws

Instances of Cocartesian must satisfy the following equations.

Left unit of +++

dimap unitL' unitL (proEmpty +++ p) === p, where unitL' = Right :: a -> Either Void a; unitL = either absurd id :: Either Void a -> a

Right unit of +++

dimap unitR' unitR (p +++ proEmpty) === p, where unitR' = Left :: a -> Either a Void; unitR = either id absurd :: Either a Void -> a

Associativity of +++

dimap assoc unassoc ((p +++ q) +++ r) === p +++ (q +++ r)

Some instance of Cocartesian optionally satisfy the following equation.

Commutativity of +++

dimap swap swap (p +++ q) === q +++ p

In this library, an instance of Cocartesian is called commutative Cocartesian if it satisfy the above commutativity law.

Instances with and without commutativity

TODO fill here: Star f, Clown Equivalence (positive) and Clown Comparison, Alternative f => Joker f (negative)

Optional laws for interaction of Cartesian and Cocartesian

If a Profunctor p is an instance of both Cartesian and Cocartesian, there are few named properties they optionally satisfy.

Left zero

dimap absurd (absurd . fst) (proEmpty *** p) === proEmpty

Left distribution

dimap undistL distL ((p +++ q) *** r) === (p *** r) +++ (q *** r), where distL, undistL are defined below

Right zero

dimap absurd (absurd . snd) (p *** proEmpty) === proEmpty

Right distribution

dimap undistR distR (r *** (p +++ q)) === (r *** p) +++ (r *** q), where distR, undistR are defined below

In this library, an instance of both Cartesian and Cocartesian is called

near-Bicartesian if it additionally satisfy Left zero and Left distribution

Bicartesian if it is near-Bicartesian, is commutative Cocartesian, and satisfy all four of Right zero and Right distribution. In other words, satisfy all of above four laws and commutativity of +++.

Instances with/without optional properties

TODO fill here

Note:

In usual mathematical abstract algebra context, it is common practice to name Right distributivity to mean Left zero and Left distibution combined, and same switching for Left distributivity.

The naming convention of this library came from Haskell convention of describing distributive properties of Alternative and MonadPlus.

https://wiki.haskell.org/index.php?title=Typeclassopedia#Laws_6

Type isomorphisms

Re-exports

class Bifunctor p => Assoc (p :: Type -> Type -> Type) where

"Semigroup-y" Bifunctors.

assoc . unassoc = id
unassoc . assoc = id
assoc . bimap (bimap f g) h = bimap f (bimap g h) . assoc

This library doesn't provide Monoidal class, with left and right unitors. Are they useful in practice?

Methods

assoc :: p (p a b) c -> p a (p b c)

unassoc :: p a (p b c) -> p (p a b) c

class Swap (p :: Type -> Type -> Type) where

Symmetric Bifunctors.

swap . swap = id

If p is a Bifunctor the following property is assumed to hold:

swap . bimap f g = bimap g f . swap

Swap isn't a subclass of Bifunctor, as for example

>>> newtype Bipredicate a b = Bipredicate (a -> b -> Bool)

is not a Bifunctor but has Swap instance

>>> instance Swap Bipredicate where swap (Bipredicate p) = Bipredicate (flip p)

Methods

swap :: p a b -> p b a

Distributivity

distL :: (Either a b, c) -> Either (a, c) (b, c)

undistL :: Either (a, c) (b, c) -> (Either a b, c)

distR :: (c, Either a b) -> Either (c, a) (c, b)

undistR :: Either (c, a) (c, b) -> (c, Either a b)