Data.Profunctor.FinFn

Documentation

data FinFn a b

makeFinFn :: forall (n :: Nat) a b. KnownNat n => (a -> Finite n) -> (Finite n -> b) -> FinFn a b

Safe construction

withFinFn :: FinFn a b -> (forall (n :: Nat). KnownNat n => (a -> Finite n) -> (Finite n -> b) -> r) -> r

Safe pattern matching

applyFinFn :: FinFn a b -> a -> b

How I would say... applyFinFn is nice, functor-ish. It commutes with Profunctor, Cartesian, Cocartesian, and composition (>>>>).

(>>>>) :: FinFn a b -> FinFn b c -> FinFn a c infixr 1

FinFn composes like Category, but can't be a Category since it does not have id.

(<<<<) :: FinFn b c -> FinFn a b -> FinFn a c infixr 1

(<<<<) = flip (>>>>)

fromMap :: Ord a => Map a b -> FinFn a (Maybe b)

applyFinFn . fromMap = flip Map.lookup