Data.Profunctor.UniconEncoding

Documentation

data Unicon a

Universal container. Any container type with finite maximum length and finite number of possible shapes can be encoded into Unicon.

Constructors

MkUnicon !Int !(Vector a)

data Builder a

A lazy version of Unicon.

This type is meant to be abstract. Only thing you can do is to convert it into Unicon using build.

build :: Builder a -> Unicon a

data Encoding a b s t

Encoding into Unicon a with partial decoding function.

Constructors

Encoding Int (s -> Builder a) (Unicon b -> Maybe (t, Vector b))

encodeWith :: Encoding a b s t -> s -> Unicon a

decodeWith :: Encoding a b s t -> Unicon b -> Maybe t

idEncoding :: Encoding a b a b