Data.Semigroup.Action

Action classes

class Semigroup s => Action s x | x -> s where

Type equipped with left semigroup action.

act must satisfy the following equation.

s1 `act` (s2 `act` x) === (s1 <> s2) `act` x

Additionally, if s is Monoid, the following equation must be satisfied too.

mempty `act` x = x

Minimal complete definition

Nothing

Methods

act :: s -> x -> x infixr 6

Left semigroup action

To implement an instance of Action, you can omit the definition of act if x is the same type to s.

default act :: s ~ x => s -> x -> x

actNtimes :: Integral b => b -> s -> x -> x

Left semigroup action repeated n times.

For positive integer n, the following should satisfy.

actNtimes n s x
 === act s $ act s $ ... n times ... $ x
 === act ('stimes' n s) x

As long as s is Monoid, actNtimes n must be defined for n = 0:

actNtimes 0 s x === act 'mempty' x === x

As long as s is Group, actNtimes n for negative n must be defined too:

actNtimes n s x === actNtimes (negate n) ('invert' s) x === act ('pow' s n)

It is allowed to define actNtimes n for n = 0 ( negative value of n) even when s is not Monoid (Group), but there is no guarantee on the result.

actNtimesMonoid :: (Integral b, Action s x, Monoid s) => b -> s -> x -> x

actNtimesGroup :: (Integral b, Action s x, Group s) => b -> s -> x -> x

class Semigroup s => RightAction s x | x -> s where

Right semigroup actions.

Minimal complete definition

Nothing

Methods

ract :: x -> s -> x infixl 7

Rihgt semigroup action

default ract :: s ~ x => x -> s -> x

ractNtimes :: Integral b => b -> x -> s -> x

Right semigroup action repeated n times.

ractNtimesMonoid :: (Integral b, RightAction s x, Monoid s) => b -> x -> s -> x

ractNtimesGroup :: (Integral b, RightAction s x, Group s) => b -> x -> s -> x

class (Group g, Action g x) => Torsor g x where

Torsor g x is a type x with free and transitive group action of g.

For an Action g x, it is called transitive if every equation

x = g `act` y

has a solution. It is called free if no such equation has multiple solution at a time. A free transitive action therefore has unique solution for every combination of x, y.

The difference x y function gives the unique solution for each x and y.

Minimal complete definition

Nothing

Methods

difference :: x -> x -> g

default difference :: g ~ x => x -> x -> g

Newtype wrappers

newtype Regular s

Constructors

Regular

Fields

getRegular :: s

newtype Mapped (f :: Type -> Type) x

Action mapped under a functor

Constructors

Mapped

Fields

getMapped :: f x

newtype Placed k v

Function type k -> v viewed as "values of v placed on locations indexed by k".

Applying semigroup actions on places k is also an action on Placed k v, except it switches left action and right action.

Constructors

Placed

Fields

getPlaced :: k -> v

newtype Pt a

Constructors

Pt a

unPt :: Pt a -> a