Data.Semigroup.Action.Class

Action classes

class Semigroup s => Action s x | x -> s where

Type equipped with left semigroup action.

act must satisfy the following equation.

s1 `act` (s2 `act` x) === (s1 '<>' s2) `act` x

Additionally, if s is a Monoid, the following equation must be satisfied too.

mempty `act` x === x

If s is a Group, it follows that

inv s1 `act` (s1 `act` x) === x
s1 `act` (inv s1 `act` x) === x

for any s1 and x.

Minimal complete definition

Nothing

Methods

act :: s -> x -> x infixr 6

Left semigroup action

To implement an instance of Action, you can omit the definition of act if x is the same type to s and the action is the left regular action i.e. act s x === s <> x.

default act :: s ~ x => s -> x -> x

actNtimes :: Integral b => b -> s -> x -> x

Left semigroup action repeated n times.

For positive integer n, the following should satisfy.

actNtimes n s x
 === act s $ act s $ ... n times ... $ x
 === act ('stimes' n s) x

As long as s is Monoid, actNtimes n must be defined for n = 0:

actNtimes 0 s x === act 'mempty' x === x

As long as s is Group, actNtimes n for negative n must be defined too:

actNtimes n s x === actNtimes (negate n) ('invert' s) x === act ('pow' s n)

It is possible that actNtimes n x for n <= 0 (n < 0) is defined and not a bottom such as undefined even when s is not a Monoid (Group). For such cases, there is no guarantee on the value of actNtimes n x.

actNtimesMonoid :: (Integral b, Action s x, Monoid s) => b -> s -> x -> x

actNtimesGroup :: (Integral b, Action s x, Group s) => b -> s -> x -> x

class Semigroup s => RightAction s x | x -> s where

Type equipped with right semigroup action

Minimal complete definition

Nothing

Methods

ract :: x -> s -> x infixl 7

Right semigroup action

default ract :: s ~ x => x -> s -> x

ractNtimes :: Integral b => b -> x -> s -> x

Right semigroup action repeated n times.

ractNtimesMonoid :: (Integral b, RightAction s x, Monoid s) => b -> x -> s -> x

ractNtimesGroup :: (Integral b, RightAction s x, Group s) => b -> x -> s -> x

class (Group g, Action g x) => Torsor g x where

Torsor g x is a type x equipped with free and transitive left action of a group g.

For an Action g x, it is called transitive if every equation

x = g `act` y

about g has a solution. It is called free if no such equation has multiple solutions at a time. A free transitive action therefore has a unique solution g for every combination of x, y.

Torsor g x instance represents the action is free and transitive by difference x y method, which gives the unique solution g of x = g act y.

Minimal complete definition

Nothing

Methods

difference :: x -> x -> g

default difference :: g ~ x => x -> x -> g

Newtype wrappers

The newtype wrappers that pick out a specific Action/RightAction derivation strategy each live in their own module, so only the wrappers you actually use are imported: Data.Semigroup.Action.Regular, Data.Semigroup.Action.Mapped, Data.Semigroup.Action.Placed, Data.Semigroup.Action.Pt.