Data.Semigroup.Action.Placed

Documentation

newtype Placed k v

Function type k -> v viewed as "values of v placed on locations indexed by k".

Applying semigroup actions on places k is also an action on Placed k v, except it switches left action and right action.

Constructors

Placed

Fields

getPlaced :: k -> v