Data.TwoMonoids.Class

Documentation

class Monoid a => TwoMonoids a where

TwoMonoids is a Monoid with another additional monoid structure.

The additional monoid is called the "additive" monoid and operators are named additively (zero, \<+\>). The original monoid is referred as "multiplicative" monoid.

Contrary to what the names suggest, TwoMonoids do not impose laws about relation between the two monoids like distributivity (like Num do.)

Laws

Monoid laws of "additive" set of operators:

(x <+> y) <+> z === x <+> (y <+> z)
zero <+> x === x
x <+> zero === x

Relation to other known algebraic structures

Near-semiring is a subclass of TwoMonoids which additionally satisfy "left distribution law" (which confusingly have another name "right distributivity").

-- in near-semiring
zero <> x  === zero
(x <+> y) <> z === (x <> z) <+> (y <> z)

Minimal complete definition

zero, (<+>) | msum

Methods

zero :: a

The unit of "additive" monoid.

(<+>) :: a -> a -> a infixr 5

The binary operator of "additive" monoid.

msum :: [a] -> a

Fold a list using <+> monoid.

mtimes :: Natural -> a -> a

n-fold repeated addition.

Beware that stimes from Data.Semigroup module is for the "multiplicative" monoid and different to mtimes.

one :: TwoMonoids a => a

Synonym of mempty but renamed to emphasize it is for "multiplicative" monoid.

mprod :: TwoMonoids a => [a] -> a

Synonym of mconcat but renamed to emphasize it is for "multiplicative" monoid.

mpow :: TwoMonoids a => Natural -> a -> a

Type-restricted synonym of mtimesDefault but renamed to emphasize it is for "multiplicative" monoid.