Data.TwoMonoids.LZ.Class

Documentation

class TwoMonoids a => LZ a where

TwoMonoids with additional property (left zero): zero is a right-absorbing element.

zero <> x = zero

Minimal complete definition

Nothing

Methods

mprodZ :: ProdZ a -> a

mprodZ is a ProdZ-algebra.

mprodZ . fmap mprodZ === mprodZ . join

When you override the default implementation, it must keep the relation to the other methods:

mprodZ Nil === one
mprodZ (Zee _) === zero
mprodZ (pure a) = mprodZ (Cons a Nil) === a

type ProdZ = ListZ ZeroEnd

ProdZ a is a formal product of a in left zero TwoMonoids, represented as ListZ ZeroEnd.

data ZeroEnd

Formal zero element.

Constructors

ZeroEnd