Data.TwoMonoids.LZLC.Class

Documentation

class LZ a => LZLC a where

TwoMonoids with left zero (LZ) and additional left catch property: one is right-absorbing element for the additive monoid.

-- left catch
one <+> x === one

Minimal complete definition

Nothing

Methods

msumZ :: SumZ a -> a

msumZ is a ZList algebra.

msumZ . fmap msumZ === msumZ . join

When you override the default implementation, it must keep the relation to the other methods:

msumZ Nil     === zero
msumZ (Zee _) === one
msumZ (pure a)  === msumZ (Cons a Nil) === a

type SumZ = ListZ OneEnd

SumZ a is a formal sum of a in left catch TwoMonoids, represented as ListZ OneEnd.

data OneEnd

Formal unit element.

Constructors

OneEnd