Data.TwoMonoids.LZLC.Free

Documentation

data Free a

The free LZLC.

Constructors

ShortExpr (Short a)

SumOf (SumZ' (Summand a))

ProdOf (ProdZ' (Factor a))

data Short a

Trivial expressions of LZLC.

Constructors

Zero

One

Lit a

type SumZ = ListZ OneEnd

SumZ a is a formal sum of a in left catch TwoMonoids, represented as ListZ OneEnd.

type SumZ' = ListZ' OneEnd

SumZ a sans trivial expressions (zero, one, pure a)

shortToSumZ :: Short a -> SumZ a

nontrivialSumZ :: SumZ a -> Either (Short a) (SumZ' a)

type ProdZ = ListZ ZeroEnd

ProdZ a is a formal product of a in left zero TwoMonoids, represented as ListZ ZeroEnd.

type ProdZ' = ListZ' ZeroEnd

ProdZ a sans trivial expressions (one, zero, pure a)

shortToProdZ :: Short a -> ProdZ a

nontrivialProdZ :: ProdZ a -> Either (Short a) (ProdZ' a)

data Summand a

Expressions which can't be written as a nontrivial sum $x <+> y$

Constructors

SummandLit a

SummandProd (ProdZ' (Factor a))

injectSummand :: Summand a -> Free a

data Factor a

Expressions which can't be written as a nontrivial product $x <> y$

Constructors

FactorLit a

FactorSum (SumZ' (Summand a))

injectFactor :: Factor a -> Free a

viewSum :: Free a -> SumZ (Summand a)

viewProd :: Free a -> ProdZ (Factor a)

interpret :: LZLC b => (a -> b) -> Free a -> b