FFunctor.Tannen

Documentation

newtype Tannen (h :: FUNCTOR) (ff :: FF) (g :: FUNCTOR) x

Wraps a FFunctor ff inside a Functor h.

This can be thought of as a functor composition below.

(Functor × Type) -[ff]-> Type -[h]-> Type

The name Tannen is an homage to Data.Bifunctor.Tannen in bifunctors package, which works in a similar way: taking Bifunctor p and Functor h, makes Bifunctor by the composition below.

(Type × Type) -[p]-> Type -[f]-> Type

Constructors

Tannen

Fields

runTannen :: h (ff g x)