FMonad.FreeT

Description

Another way to make FreeT an instance of FMonad

FreeT can be FMonad in two different ways. There is already an instance:

instance Functor f => FMonad (FreeT f) where
  fpure :: Functor m => m ~> FreeT f m
  fbind :: (Functor m, Functor n) => (m ~> FreeT f n) -> (FreeT f m ~> FreeT f n)

In addition to this standard instance, FreeT f m have FMonad-like structure by treating f as the parameter while fixing m to some arbitrary Monad.

fpureFst :: (Monad m) => (Functor f) => f ~> FreeT f m
fbindFst :: (Monad m) => (Functor f, Functor g) => (f ~> FreeT g m) -> (FreeT f m ~> FreeT g m)

This module provides a newtype wrapper FreeT' to use these as a real FMonad instance.

Documentation

newtype FreeT' (m :: Type -> Type) (f :: Type -> Type) b

FreeT' is a FreeT, but with the order of its arguments flipped.

FreeT' m f a ≡ FreeT f m a

Constructors

WrapFreeT'

Fields

unwrapFreeT' :: FreeT f m b

liftM' :: forall m a (f :: Type -> Type). Functor m => m a -> FreeT' m f a

Lift of the Monad side.

fpureFst :: forall (m :: Type -> Type) (f :: Type -> Type). (Monad m, Functor f) => f ~> FreeT f m

fpure to the first parameter of FreeT

fbindFst :: forall (m :: Type -> Type) (f :: Type -> Type) (g :: Type -> Type). (Monad m, Functor f, Functor g) => (f ~> FreeT g m) -> FreeT f m ~> FreeT g m

fbind to the first parameter of FreeT