GHC.TypeNats.Extra

Documentation

data SNatView (n :: Natural) where

A view of SNat as if it is defined inductively.

Constructors

ViewZero :: SNatView 0

ViewSucc :: forall (n1 :: Natural). SNat n1 -> SNatView (n1 + 1)

viewSNat :: forall (n :: Nat). SNat n -> SNatView n

Perform case analysis on SNat.

data SNat (n :: Nat) where

A value-level witness for a type-level natural number. This is commonly referred to as a singleton type, as for each n, there is a single value that inhabits the type SNat n (aside from bottom).

The definition of SNat is intentionally left abstract. To obtain an SNat value, use one of the following:

The natSing method of KnownNat.

The SNat pattern synonym.

The withSomeSNat function, which creates an SNat from a Natural number.

Since: base-4.18.0.0

Bundled Patterns

pattern SNat :: () => KnownNat n => SNat n
Wrapper for GHC.TypeNats.SNat

pattern Zero :: () => n ~ 0 => SNat n
Pattern synonym Zero and Succ treats SNat as if it was inductively defined type.

@ -- Virtual definition of SNat data SNat n where Zero :: SNat 0 Succ :: SNat n -> SNat (n + 1)

pattern Succ :: forall n m. () => n ~ (m + 1) => SNat m -> SNat n

(%+) :: forall (n :: Nat) (m :: Nat). SNat n -> SNat m -> SNat (n + m)

(%-) :: forall (n :: Nat) (m :: Nat). SNat n -> SNat m -> SNat (n - m)

(%*) :: forall (n :: Nat) (m :: Nat). SNat n -> SNat m -> SNat (n * m)

(%^) :: forall (n :: Nat) (m :: Nat). SNat n -> SNat m -> SNat (n ^ m)