ModelFinder.Examples.Monad

Documentation

data M

Constructors

M0

M1

data I

Constructors

I0

I1

data MonadSig x where

Signature of data describing Monad on T a = (M, I -> a)

Let T a be a functor of the following shape.

type T a = (M, I -> a)

Any Monad on type can be written as follows:

pure :: a -> T a
pure a = (unit, const a)

join :: T (T a) -> T a
join (m, f) = (munge m v, \i -> v (index1 m v i) (index2 m v i))
  where v = fst . f
        h = snd . f

using these unknown values and functions.

unit :: M
munge :: M -> (I -> M) -> M
index1 :: M -> (I -> M) -> I -> I
index2 :: M -> (I -> M) -> I -> I

By translating the Monad laws written in terms of (pure, join) to that of (unit, munge, index1, index2), we can make equational theory of Monads on T.

MonadSig is the signature of terms, and equations are implemented afterwards.

Constructors

Unit :: MonadSig M

Munge :: M -> (I :-> M) -> MonadSig M

Index1 :: M -> (I :-> M) -> I -> MonadSig I

Index2 :: M -> (I :-> M) -> I -> MonadSig I

newtype a :-> b

Function with finite domain can be treated as data (compare two functions, show to text, ...)

Constructors

Fn

Fields

($$) :: a -> b

searchMonad :: () -> [Solution MonadSig]