ModelFinder.Model

Documentation

class Model k a model | model -> k a where

model represents possible assignments φ :: k -> a.

An assignment can be partial

A model can represent a priori knowledge about the assignments (e.g. injectivity "type" restriction naturality / etc.)

Minimal complete definition

guess, guessMany, enterDef

Methods

guess :: k -> model -> [a]

Get the current possible assignments. Returning singleton means the value of k is already settled. It can be empty if the model has gotten inconsistent.

guessMany :: NonEmpty k -> model -> [a]

Get the intersection of guesses for the given multiple queries.

enterDef :: [k] -> a -> model -> Maybe (model, [(k, a)])

Update the model accoring to newly added definitive equality φ(k1) = φ(k2) = ... = y :: a. Returns the updated model and newly discovered definitive equalities. Returns Nothing if the given equality is incompatible with model.

enterEqs :: [k] -> model -> Maybe (model, [(k, a)])

Update the model accoring to newly added equality φ(k1) = φ(k2) = ... = φ(kn). Returns the updated model and newly discovered definitive equalities. Returns Nothing if the given equality is incompatible with model.

newtype DumbModel k a

DumbModel remember nothing and just returns anything possible as guess.

Constructors

DumbModel

Fields

dumbUniverse :: [a]

newDumbModel :: [a] -> DumbModel k a

data SimpleModel k a

SimpleModel remembers a value for each specific f a. The guess returns set of arbitrary a unless the requested f a value is remembered.

Constructors

SimpleModel

Fields

simpleUniverse :: [a]

simpleGuesses :: !(Map k a)

newSimpleModel :: [a] -> SimpleModel k a

newtype WrapperModel k' k a model

Constructors

WrapperModel

Fields

unwrapModel :: model