TrailAsAdjunction

Description

Explain Trail monad in terms of adjoint functors

Building blocks

newtype Const a (b :: k)

The Const functor.

Constructors

Const

Fields

getConst :: a

type Pair = (,)

Pair :: Type -> Functor

newtype Shape (f :: Type -> Type)

Shape :: Functor -> Type

Constructors

Shape

Fields

getShape :: f ()

newtype Emptied (f :: Type -> Type)

Emptied :: Functor -> Type

Constructors

Emptied

Fields

getEmptied :: f Void

type Y = (->)

Yoneda embedding

Y :: Type^op -> Functor

newtype Γ (f :: Type -> Type)

Γ :: Functor -> Type^op

Constructors

Γ

Fields

getΓ :: forall x. f x -> x

mapPair :: (a -> b) -> Pair a ~> Pair b

mapShape :: forall (f :: Type -> Type) (g :: Type -> Type). (f ~> g) -> Shape f -> Shape g

mapEmptied :: forall (f :: Type -> Type) (g :: Type -> Type). (f ~> g) -> Emptied f -> Emptied g

mapY :: (a -> b) -> Y b ~> Y a

mapΓ :: forall (f :: Type -> Type) (g :: Type -> Type). (f ~> g) -> Γ g -> Γ f

Adjunctions

There is a chain of three adjunctions between Type and Functor

                   Type
  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     |            ^          |            ^
     |            |          |            |
Pair |  -|  Shape | -| Const | -| Emptied | 
     |            |          |            |
     v            |          v            |
  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
                  Functor

And an adjunction between Type^op and Functor

      Type^op
  ~~~~~~~~~~~~~~~~
     ^      |
     |      |
   Γ |  -|  | Y
     |      |
     |      v
  ~~~~~~~~~~~~~~~~
     Functor

adjunctPairShape :: forall a (f :: Type -> Type). (Pair a ~> f) -> a -> Shape f

adjunctPairShape' :: forall (f :: Type -> Type) a. Functor f => (a -> Shape f) -> Pair a ~> f

adjunctShapeConst :: forall (f :: Type -> Type) a. Functor f => (Shape f -> a) -> f ~> (Const a :: Type -> Type)

adjunctShapeConst' :: forall (f :: Type -> Type) a. (f ~> (Const a :: Type -> Type)) -> Shape f -> a

adjunctConstEmptied :: forall a (f :: Type -> Type). ((Const a :: Type -> Type) ~> f) -> a -> Emptied f

adjunctConstEmptied' :: forall (f :: Type -> Type) a. Functor f => (a -> Emptied f) -> (Const a :: Type -> Type) ~> f

adjunctΓY :: forall a (f :: Type -> Type). (a -> Γ f) -> f ~> Y a

adjunctΓY :: (Γ f `Op' a) -> (f ~> Y a)

adjunctΓY' :: forall (f :: Type -> Type) a. (f ~> Y a) -> a -> Γ f

adjunctΓY' :: (f ~> Y a) -> (Γ f Op a)

Monads and Comonads from these adjunctions

newtype J0 (mm :: (Type -> Type) -> Type -> Type) a

J0 mm ~ Emptied ∘ mm ∘ Const

Constructors

J0

Fields

runJ0 :: mm (Const a :: Type -> Type) Void

newtype K1 (ww :: (Type -> Type) -> Type -> Type) a

K ww ~ Shape ∘ ww ∘ Const

Constructors

K1

Fields

runK1 :: ww (Const a :: Type -> Type) ()

newtype J1 (mm :: (Type -> Type) -> Type -> Type) a

J1 mm ~ Emptied ∘ mm ∘ Const

Also, J1 ~ Trail

Constructors

J1

Fields

runJ1 :: mm (Pair a) ()

newtype RR (ww :: (Type -> Type) -> Type -> Type) a

RR ww ~ Γ ∘ ww ∘ Y

Constructors

RR

Fields

runRR :: Γ (ww (Y a))

Examples

J0 and J1 turn an FMonad mm to a Monad. The following table shows what is the Monad they make for each FMonad.

mm	Compose m	Precompose m	Sum f	Free	FreeT' m

J0 mm	m	Identity	Either (f Void)	Identity	m

J1 mm	m	Writer (Ap m ())	Either (f ())	[]	ListT m

K1 takes an FComonad ww and turns them a Comonad. RR takes an FComonad ww and turns them a Monad, instead of Comonad.

ww	Compose w	Precompose w	Product f	Cofree	Sum ff gg

K1 ww	w	Identity	Env (f ())	Identity	K1 ff :+: K1 gg

RR ww	Co w	R w	Exp1 f Identity	[]	??

Where

Co is a Monad defined in Control.Monad.Co from kan-extensions package

R is a Monad defined below

Exp1 is defined in Data.Functor.Exp

newtype R (w :: Type -> Type) a

RR ww specialized to ww = Precompose w

Constructors

R

Fields

runR :: forall r. (a -> w r) -> r