Control.Alternative.Free.LZLC

Description

Free alternative (with left zero + left catch).

Alternative laws were not set to one, clear definition, but there are two major ones.

For an instance of (Alternative f), both laws have these in common.

Inherited laws from Applicative

(empty, <|>) forms monoid on f a for any type a.

Left zero law: empty <*> x === empty.

Candidate #1 of the Alternative law have Left distribution law.

-- Left distribution
(x <|> y) <*> z === (x <*> z) <|> (y <*> z)

Another candidate #2 have Left catch law instead.

-- Left catch
pure x <|> y === pure x

Reference Typeclassopedia https://wiki.haskell.org/Typeclassopedia#Laws_6 for more about these laws.

The "free alternative" construction for the alternative #1 (with Left distribution) is known and implemented.

https://people.cs.kuleuven.be/~tom.schrijvers/Research/talks/ppdp2015.pdf

https://hackage.haskell.org/package/free-5.2/docs/Control-Alternative-Free.html

This module provides the free alternative #2.

Documentation

type Az = Free

data Free (f :: Type -> Type) a where

Constructors

Pure :: forall a (f :: Type -> Type). a -> Free f a

Zero :: forall (f :: Type -> Type) a. Free f a

Lift :: forall (f :: Type -> Type) a. f a -> Free f a

SumOf :: forall (f :: Type -> Type) a. Sz' (SummandF f) a -> Free f a

ApOf :: forall (f :: Type -> Type) a. Az' (FactorF f) a -> Free f a

data SummandF (f :: Type -> Type) a where

Constructors

SummandLift :: forall (f :: Type -> Type) a. f a -> SummandF f a

SummandAp :: forall (f :: Type -> Type) a. Az' (FactorF f) a -> SummandF f a

data FactorF (f :: Type -> Type) a where

Constructors

FactorLift :: forall (f :: Type -> Type) a. f a -> FactorF f a

FactorSum :: forall (f :: Type -> Type) a. Sz' (SummandF f) a -> FactorF f a

data Sz (f :: Type -> Type) a

Constructors

Nend

Zend a

Cons (f a) (Sz f a)

data Sz' (f :: Type -> Type) a

Constructors

SzFz (f a) a

SzLong (f a) (f a) (Sz f a)

data Az' (f :: Type -> Type) a where

Constructors

AzFz :: forall (f :: Type -> Type) a1 a. f a1 -> Az' f a

AzLong :: forall (f :: Type -> Type) a1 b a. f a1 -> f b -> Az f (b -> a1 -> a) -> Az' f a

injectSummand :: forall (f :: Type -> Type) a. SummandF f a -> Free f a

injectFactor :: forall (f :: Type -> Type) a. FactorF f a -> Free f a

viewSum :: forall (f :: Type -> Type) a. Free f a -> Sz (SummandF f) a

viewAp :: forall (f :: Type -> Type) a. Free f a -> Az (FactorF f) a