Control.Alternative.Free.LZLC

Description

Free alternative (with left zero + left catch).

Alternative laws were not set to one, clear definition, but there are two major ones.

For an instance of (Alternative f), both laws have these in common.

Inherited laws from Applicative

(f a, empty, <|>) forms monoid for any type a.

Left zero law: empty <*> x === empty.

Candidate #1 of the Alternative law have Left distribution law.

-- Left distribution
(x <|> y) <*> z === (x <*> z) <|> (y <*> z)

Another candidate #2 have Left catch law instead.

-- Left catch
pure x <|> y === pure x

Reference Typeclassopedia https://wiki.haskell.org/Typeclassopedia#Laws_6 for more about these laws.

The "free alternative" construction for the alternative #1 (with Left distribution) is known and implemented.

https://people.cs.kuleuven.be/~tom.schrijvers/Research/talks/ppdp2015.pdf

https://hackage.haskell.org/package/free-5.2/docs/Control-Alternative-Free.html

This module provides the free alternative #2.

Type definitions

data Free (f :: Type -> Type) a where

The Free (left zero + left catch) Alternative.

The constructors of Free is named intentionally verbose.

To construct a value of Free f a, a convenient and recommended way is to use liftFree and methods of Alternative (Free f) instances.

To pattern match a value of Free f a, a convenient and recommended way is to use ApZOf and SumZOf pattern synonyms, along with viewFactor and viewSummand to pattern match on Factor and Summand further.

Constructors

FreeTrivial :: forall (f :: Type -> Type) a. Trivial f a -> Free f a

FreeSumZOf' :: forall (f :: Type -> Type) a. NontrivialSumZ (Summand f) a -> Free f a

FreeApZOf' :: forall (f :: Type -> Type) a. NontrivialApZ (Factor f) a -> Free f a

Bundled Patterns

pattern SumZOf :: SumZ (Summand f) a -> Free f a
View Free f a as a sum of Summand f, terminated by either Nil or Zee a (corresponding empty or pure a respectively)

pattern ApZOf :: Functor f => ApZ (Factor f) a -> Free f a
View Free f a as a product of Factor f, terminated by either Zero or Pure a (corresponding empty or pure a respectively)

newtype Summand (f :: Type -> Type) a

Subexpressions of Free f a which can't be written as any one of

empty

pure a

Nontrivial sum x | y

Constructors

Summand

Fields

runSummand :: Either (f a) (NontrivialApZ (Factor f) a)

viewSummand :: Summand f a -> Either (f a) (ApZ (Factor f) a)

Summand f a is either f a or a product of multiple Factor f values. The Right case never returns a trivial summand, which is one of

Zero,

Pure a,

liftApZ fac for one fac :: Factor f a

newtype Factor (f :: Type -> Type) a

Subexpressions of Free f a which can't be written as any one of

empty

pure a

Nontrivial product x * y

Constructors

Factor

Fields

runFactor :: Either (f a) (NontrivialSumZ (Summand f) a)

viewFactor :: Factor f a -> Either (f a) (SumZ (Summand f) a)

Factor f a is either f a or a sum of multiple Summand f values. The Right case never returns a trivial factor, which is one of

Nil,

Zee a,

Cons suma Nil for one suma :: Summand f a

viewSumZ :: forall (f :: Type -> Type) a. Free f a -> SumZ (Summand f) a

reviewSumZ :: forall (f :: Type -> Type) a. SumZ (Summand f) a -> Free f a

viewApZ :: forall (f :: Type -> Type) a. Free f a -> ApZ (Factor f) a

reviewApZ :: forall (f :: Type -> Type) a. Functor f => ApZ (Factor f) a -> Free f a

Universal properties

hoistFree :: (forall x. f x -> g x) -> Free f a -> Free g a

liftFree :: f a -> Free f a

foldFree :: forall f g a. Alternative g => (forall x. f x -> g x) -> Free f a -> g a

Auxiliary definitions

data Trivial (f :: Type -> Type) a

Trivial expressions

Constructors

TrivialZero

TrivialPure a

TrivialLift (f a)

type SumZ (f :: Type -> Type) a = ListZ a (f a)

Formal sum of f a ending in either pure a (Zee a) or empty (Nil).

data NontrivialSumZ (f :: Type -> Type) a

Trivial f a + NontrivialSumZ f a ~ SumZ f a = ListZ a (f a)

Constructors

ConsZee (f a) a

ConsMany (f a) (f a) (ListZ a (f a))

data NontrivialApZ (f :: Type -> Type) a where

Trivial f a + NontrivialApZ f a ~ ApZ f a

Constructors

ApZero :: forall (f :: Type -> Type) a1 a. f a1 -> NontrivialApZ f a

ApMany :: forall (f :: Type -> Type) a1 b a. f a1 -> f b -> ApZ f (b -> a1 -> a) -> NontrivialApZ f a