Data.Profunctor.Cartesian

Documentation

class Profunctor p => Cartesian (p :: Type -> Type -> Type) where

Minimal complete definition

proUnit, (proProduct | (***) | (&&&))

Methods

proUnit :: p a ()

Unit of the product.

The type of proUnit can be understood as proUnit' :: p () () treated to have more generally-typed p a () using lmap.

const () :: a -> ()
proUnit' :: p () ()
lmap (const ()) proUnit' :: p a ()

proProduct :: (a -> (a1, a2)) -> ((b1, b2) -> b) -> p a1 b1 -> p a2 b2 -> p a b

(***) :: p a b -> p a' b' -> p (a, a') (b, b')

Product of two profunctors

(&&&) :: p a b -> p a b' -> p a (b, b')

Alternative way to define the product (pronounced "fan-out")

proPower :: forall (n :: Nat) a b. KnownNat n => p a b -> p (Finite n -> a) (Finite n -> b)

Function from finite types can be constructed as iterated product.

There is a default implementaion, but it can be more efficient implementation.

prodDay :: forall (r :: Type -> Type -> Type) (p :: Type -> Type -> Type) (q :: Type -> Type -> Type). Cartesian r => (p :-> r) -> (q :-> r) -> Day (,) p q :-> r

class Profunctor p => Cocartesian (p :: Type -> Type -> Type) where

Minimal complete definition

proEmpty, (proSum | (+++) | (|||))

Methods

proEmpty :: p Void b

Unit of the sum.

The type of proEmpty can be understood as proEmpty' :: p Void Void treated to have more generally-typed p Void b using rmap.

absurd    :: Void -> b
proEmpty' :: p Void Void
rmap absurd proEmpty' :: p Void b

proSum :: (a -> Either a1 a2) -> (Either b1 b2 -> b) -> p a1 b1 -> p a2 b2 -> p a b

(+++) :: p a b -> p a' b' -> p (Either a a') (Either b b')

Sum of two profunctors

(|||) :: p a b -> p a' b -> p (Either a a') b

Alternative way to define the sum (pronounced "fan-in")

proTimes :: forall (n :: Nat) a b. KnownNat n => p a b -> p (Finite n, a) (Finite n, b)

Pairing with finite types can be constructed as iterated sum.

There is a default implementaion, but it can be more efficient implementation.

sumDay :: forall (r :: Type -> Type -> Type) (p :: Type -> Type -> Type) (q :: Type -> Type -> Type). Cocartesian r => (p :-> r) -> (q :-> r) -> Day Either p q :-> r

Cartesian p iff forall x. Applicative (p x)

pureDefault :: Cartesian p => b -> p a b

proUnitDefault :: Applicative (p a) => p a ()

liftA2Default :: Cartesian p => (a -> b -> c) -> p x a -> p x b -> p x c

fanoutDefault :: (Profunctor p, Applicative (p a)) => p a b1 -> p a b2 -> p a (b1, b2)

Utilities

describeFinite :: forall (n :: Nat) p. (KnownNat n, Cartesian p, Cocartesian p) => p (Finite n) (Finite n)

describeFiniteBits :: (FiniteBits a, Cartesian p, Cocartesian p) => p a a

About distributivy laws between Cartesian and Cocartesian