Data.Profunctor.UniconEncoding

Documentation

data Unicon a

Universal container. Any container type with finite maximum length and finite number of possible shapes can be encoded into Unicon.

Constructors

MkUnicon !Int [a]

data Encoding a s t

Encoding into Unicon a with partial decoding function.

Constructors

Encoding Int (s -> Unicon a) (Unicon a -> Maybe (t, [a]))

encodeId :: Encoding a a a